requirements.txt
aiogram==3.4.1
pytz==2024.1
python-dotenv==1.0.1
PyYAML==6.0.1
cachetools==5.4.0